Change id's from long to int.

+++ browser/backend/app/src/main/java/com/example/backend/entities/Food.java
@@ -1,11 +1,7 @@
 package com.example.backend.entities;
+import jakarta.persistence.*;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
 import lombok.Data;
 
 @Entity
@@ -15,7 +11,7 @@ public class Food implements BaseEntity<Food> {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String name; // NOT NULL
     private String barcode;
